+++ src/frontend/apps/lti_site/apps/markdown/components/MarkdownWizard/index.spec.tsx
@@ -111,7 +111,7 @@ describe('MarkdownWizard', () => {
       'Mon titre',
     );
     await userEvent.click(
-      screen.getByRole('button', { name: /Select language/i }),
+      screen.getByRole('combobox', { name: /Select language/i }),
     );
     await userEvent.click(
       await screen.findByRole('option', { name: /French/i }),