+++ src/frontend/packages/lib_video/src/components/common/VideoWidgetProvider/widgets/LiveJoinMode/index.tsx
@@ -1,4 +1,5 @@
-import { Box, Select } from 'grommet';
+import { Select } from '@openfun/cunningham-react';
+import { Box } from 'grommet';
 import { FoldableItem, JoinMode, report } from 'lib-components';
 import React from 'react';
 import { toast } from 'react-hot-toast';
@@ -22,9 +23,14 @@ const messages = defineMessages({
     id: 'components.LiveJoinMode.title',
   },
   selectLabel: {
-    defaultMessage: 'Select join the discussion mode',
+    defaultMessage: 'Join the discussion mode',
     description: 'The label for the select to set join modes.',
     id: 'components.LiveJoinMode.selectLabel',
+  },
+  selectLabelInfo: {
+    defaultMessage: 'Choose the mode when someone join the discussion',
+    description: 'The text under the select to set join modes.',
+    id: 'components.LiveJoinMode.selectLabelInfo',
   },
   approval: {
     defaultMessage: 'Accept joining the discussion after approval',
@@ -95,20 +101,19 @@ export const LiveJoinMode = () => {
       <Box direction="column" gap="small">
         <Select
           aria-label={intl.formatMessage(messages.selectLabel)}
+          label={intl.formatMessage(messages.selectLabel)}
           options={options}
-          labelKey="label"
-          replace={false}
-          valueKey={{ key: 'value', reduce: true }}
+          fullWidth
           value={video.join_mode}
-          onChange={({
-            option,
-          }: {
-            option: { label: string; value: JoinMode };
-          }) => {
-            videoMutation.mutate({
-              join_mode: option.value,
-            });
+          onChange={(evt) => {
+            if (evt.target.value !== video.join_mode) {
+              videoMutation.mutate({
+                join_mode: evt.target.value as JoinMode,
+              });
+            }
           }}
+          clearable={false}
+          text={intl.formatMessage(messages.selectLabelInfo)}
         />
       </Box>
     </FoldableItem>