+++ src/frontend/packages/lib_video/src/components/common/VideoWizard/CreateVOD/index.tsx
@@ -181,6 +181,7 @@ export const CreateVOD = ({
               })
             }
             value={wizardedVideo.title || ''}
+            fullWidth
           />
 
           <UploadVideoForm